+++ src/dexcontrol/__init__.py
@@ -15,8 +15,6 @@ This package provides interfaces for controlling and monitoring robot systems.
 It serves as the primary API for interacting with Dexmate robots.
 """
 
-from __future__ import annotations
-
 import os
 from pathlib import Path
 from typing import Final
@@ -27,9 +25,13 @@ from rich.logging import RichHandler
 # DO NOT REMOVE this following import, it is needed for hydra to find the config
 import dexcontrol.config  # pylint: disable=unused-import
 from dexcontrol.exceptions import (
+    ComponentError,
+    ComponentNotAvailableError,
     ConfigurationError,
     DexcontrolError,
+    ModelNotSupportedError,
     RobotConnectionError,
+    SensorNotAvailableError,
     ServiceUnavailableError,
 )
 from dexcontrol.robot import Robot
@@ -38,7 +40,7 @@ from dexcontrol.utils.constants import COMM_CFG_PATH_ENV_VAR
 # Package-level constants
 LIB_PATH: Final[Path] = Path(__file__).resolve().parent
 CFG_PATH: Final[Path] = LIB_PATH / "config"
-MIN_SOC_SOFTWARE_VERSION: int = 360
+MIN_SOC_SOFTWARE_VERSION: int = 419
 
 logger.configure(
     handlers=[
@@ -86,6 +88,10 @@ __all__ = [
     "ROBOT_CFG_PATH",
     # Exceptions
     "DexcontrolError",
+    "ComponentError",
+    "ComponentNotAvailableError",
+    "ModelNotSupportedError",
+    "SensorNotAvailableError",
     "ConfigurationError",
     "RobotConnectionError",
     "ServiceUnavailableError",